First refactoring to discharge a bit the Resolver. (#1711)
* First refactoring to discharge a bit the Resolver.
* TypeConstraint is now a record
* Fixed warnings
* Run formatter
